gcp/requirements.txt
fastapi
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv==1.0.0
boto3
